Moved services folder out of components
